Adopt coroutine library 'co' for flow control
